fix:bug and filtrage

+++ src/main/kotlin/server/web/casa/app/user/domain/model/TypeAccount.kt
@@ -1,6 +1,6 @@
 package server.web.casa.app.user.domain.model
 
 data class TypeAccount(
-    val typeAccountId : Int,
-    val name : String
+    val typeAccountId: Long,
+    val name: String
 )
